+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ### 0.1.0 (Next)
 
+* [#1](https://github.com/dblock/danger-changelog/pull/1): Added `have_you_updated_changelog?`, have you updated CHANGELOG.md? - [@dblock](https://github.com/dblock).
 * Your contribution here.